fix(popup): show surface toggle in non-connected views too
Once a user enables side-panel/sidebar mode, the toolbar click goes to
the persistent surface forever. If they then land in NeedsInstall,
NeedsLogin, NeedsUpdate, or Disconnected, those views never rendered
the toggle, so there was no way back from the UI.

Extract renderUiSurfaceRow into its own component module along with a
renderUiSurfaceFooter wrapper, and call the wrapper from each non-
connected view. The connected view keeps using the bare row inside its
quick-settings block.
