+++ packages/shared/src/popup/components/ui-surface-row.test.ts
@@ -1,7 +1,7 @@
 // @vitest-environment happy-dom
 import { describe, it, expect, vi, beforeEach } from "vitest";
 import { writeUiSurface } from "../../background/ui-surface";
-import { renderUiSurfaceRow } from "./connected";
+import { renderUiSurfaceFooter, renderUiSurfaceRow } from "./ui-surface-row";
 
 vi.mock("../../background/ui-surface", () => ({
   readUiSurface: vi.fn(() => Promise.resolve("popup" as const)),
@@ -29,5 +29,13 @@ describe("Open as side panel toggle", () => {
     checkbox.checked = true;
     checkbox.dispatchEvent(new Event("change"));
     expect(writeUiSurface).toHaveBeenCalledWith("sidePanel");
+  });
+
+  it("footer wraps the row in a .ui-surface-footer container appended to the view", () => {
+    const view = document.createElement("div");
+    renderUiSurfaceFooter(view);
+    const footer = view.querySelector(".ui-surface-footer");
+    expect(footer).not.toBeNull();
+    expect(footer?.parentElement).toBe(view);
   });
 });
