+++ packages/shared/src/popup/views/connected.ts
@@ -10,7 +10,7 @@ import { createToggle } from "../components/toggle-switch";
 import { renderExitNodes } from "./exit-nodes";
 import { renderProfiles } from "./profiles";
 import { iconChevronRight } from "../icons";
-import { readUiSurface, writeUiSurface, type UiSurface } from "../../background/ui-surface";
+import { renderUiSurfaceRow } from "../components/ui-surface-row";
 
 type SubViewRenderer = (root: HTMLElement, state: TailscaleState, onBack: () => void) => void;
 
@@ -432,24 +432,6 @@ export function renderConnected(root: HTMLElement, state: TailscaleState): void 
   shell.appendChild(detail);
 
   root.appendChild(shell);
-}
-
-export async function renderUiSurfaceRow(parent: HTMLElement): Promise<void> {
-  const current = await readUiSurface();
-  const row = document.createElement("div");
-  row.className = "setting-row";
-
-  const label = document.createElement("span");
-  label.className = "setting-label";
-  label.textContent = "Open as side panel";
-  row.appendChild(label);
-
-  const toggle = createToggle(current === "sidePanel", (checked) => {
-    const next: UiSurface = checked ? "sidePanel" : "popup";
-    void writeUiSurface(next);
-  });
-  row.appendChild(toggle);
-  parent.appendChild(row);
 }
 
 // Track health key to detect changes during in-place updates
